chore: add tests, split documents diffing/merging tests into several test files
